+++ scripts/bridges/il2cpp-0.12.2.js
@@ -1,0 +1,3 @@
+'use strict';
+
+import 'frida-il2cpp-bridge-0-12-2';