+++ scripts/bridges/il2cpp-0.13.0.js
@@ -1,0 +1,3 @@
+'use strict';
+
+import 'frida-il2cpp-bridge-0-13-0';